Rename ecu_variant_matcher.py module
